# packages/shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghar Khoj - Broker-Free Room Rental Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /verification:
    post:
      operationId: submitVerification
      tags: [verification]
      summary: Submit user verification documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitVerificationInput"
      responses:
        "201":
          description: Verification submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDoc"
  /verification/{userId}/approve:
    post:
      operationId: approveVerification
      tags: [verification]
      summary: Admin approve user verification
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verification approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationDoc"
...
components:
  schemas:
...
    VerificationDoc:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        docType:
          type: string
          enum: [citizenship, nid, passport, license]
        docUrl:
          type: string
        selfieUrl:
          type: string
        citizenshipNumber:
          type: string
        fullNameCitizenship:
          type: string
        dateOfBirth:
          type: string
        issueDate:
          type: string
        docPhotoUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNote:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, userId, docType, status, createdAt]
...
    SubmitVerificationInput:
      type: object
      properties:
        userId:
          type: integer
        docType:
          type: string
          enum: [citizenship, nid, passport, license]
        docUrl:
          type: string
        selfieUrl:
          type: string
        citizenshipNumber:
          type: string
        fullNameCitizenship:
          type: string
        dateOfBirth:
          type: string
        issueDate:
          type: string
        docPhotoUrl:
          type: string
      required: [userId, docType]
...